+++ tests/sample_sdk.py
@@ -1,5 +1,5 @@
 from sdklib.http import HttpSdk
-from sdklib.http.renderers import MultiPartRender, FormRender
+from sdklib.http.renderers import MultiPartRenderer, FormRenderer
 from sdklib.util.parser import parse_args
 
 
@@ -24,7 +24,7 @@ class SampleHttpSdk(HttpSdk):
         :return: SdkResponse
         """
         params = parse_args(name=name, description=description, city=city)
-        return self._http_request("POST", self.API_RESTAURANTS_URL_PATH, body_params=params, render=FormRender())
+        return self._http_request("POST", self.API_RESTAURANTS_URL_PATH, body_params=params, render=FormRenderer())
 
     def update_restaurant(self, name, main_image, description=None, city=None):
         """
@@ -34,4 +34,4 @@ class SampleHttpSdk(HttpSdk):
         params = parse_args(name=name, description=description, city=city)
         files = parse_args(mainImage=main_image)
         return self._http_request("PUT", self.API_RESTAURANTS_URL_PATH, body_params=params, files=files,
-                                  render=MultiPartRender())
+                                  render=MultiPartRenderer())